Changed folder strcutre
