+++ src/auth/services/auth.service.ts
@@ -1,5 +1,5 @@
 import { Injectable } from '@nestjs/common';
-import { UsersService } from '../users/users.service';
+import { UsersService } from '../../users/services/users.service';
 
 @Injectable()
 export class AuthService {
