+++ src/users/controllers/users.controller.ts
@@ -1,0 +1,37 @@
+import {
+  ClassSerializerInterceptor,
+  Controller,
+  Get,
+  HttpException,
+  HttpStatus,
+  Param,
+  UseInterceptors,
+} from '@nestjs/common';
+import { UsersService } from '../services/users.service';
+import { SerializedUser } from '../../types';
+import { ParseIntPipe } from '@nestjs/common';
+@Controller('users')
+export class UserController {
+  constructor(private readonly usersService: UsersService) {}
+
+  @UseInterceptors(ClassSerializerInterceptor)
+  @Get('')
+  getUsers() {
+    return this.usersService.getUsers();
+  }
+
+  @UseInterceptors(ClassSerializerInterceptor)
+  @Get('/username/:username')
+  getByUsername(@Param('username') username: string) {
+    const user = this.usersService.getUserByUsername(username);
+    if (user) return new SerializedUser(user);
+    else throw new HttpException('User not found', HttpStatus.BAD_REQUEST);
+  }
+
+  @UseInterceptors(ClassSerializerInterceptor)
+  @Get('id/:id')
+  getById(@Param('id', ParseIntPipe) id: number) {
+    const user = this.usersService.getUserById(id);
+    if (user) return new SerializedUser(user);
+  }
+}
