--- src/users/users.controller.ts
@@ -1,32 +1,0 @@
-import {
-  ClassSerializerInterceptor,
-  Controller,
-  Get,
-  HttpException,
-  HttpStatus,
-  Param,
-  UseInterceptors,
-} from '@nestjs/common';
-import { UsersService } from './users.service';
-import { SerializedUser } from '../utils/types';
-
-@Controller('users')
-export class UserController {
-  constructor(private readonly usersService: UsersService) {}
-
-  @UseInterceptors(ClassSerializerInterceptor)
-  @Get('')
-  getUsers() {
-    return this.usersService.getUsers();
-  }
-
-  @UseInterceptors(ClassSerializerInterceptor)
-  @Get('/:username')
-  getUserByUsername(@Param('username') username: string) {
-    const user = this.usersService.getUserByUsername(username);
-
-    if (!user)
-      throw new HttpException('User not found', HttpStatus.BAD_REQUEST);
-    return new SerializedUser(user);
-  }
-}
