+++ src/users/users.module.ts
@@ -1,6 +1,6 @@
 import { Module } from '@nestjs/common';
-import { UsersService } from './users.service';
-import { UserController } from './users.controller';
+import { UsersService } from './services/users.service';
+import { UserController } from './controllers/users.controller';
 
 @Module({
   controllers: [UserController],
